Created another example in the routers_part2.dat file
Keep in mind that we are using simple static routing in these examples. There are no fancy algorithms.

This is made apparent with the "EXTRA EXAMPLE" where we show an inefficient scenerio when Host B is trying to Communicate with Host C. The inefficiency comes from there being a "Default Gate"

I also fixed a minor MAC Address mistake in the file